+++ src/Opulence/Ioc/Container.php
@@ -4,7 +4,7 @@
  * Opulence
  *
  * @link      https://www.opulencephp.com
- * @copyright Copyright (C) 2017 David Young
+ * @copyright Copyright (C) 2019 David Young
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
 
@@ -27,7 +27,7 @@ class Container implements IContainer
     protected $currentTarget;
     /** @var array The stack of targets */
     protected $targetStack = [];
-    /** @var IBinding[][] The list of bindings */
+    /** @var IContainerBinding[][] The list of bindings */
     protected $bindings = [];
     /** @var array The cache of reflection constructors and their parameters */
     protected $constructorReflectionCache = [];
@@ -45,7 +45,7 @@ class Container implements IContainer
      */
     public function bindFactory($interfaces, callable $factory, bool $resolveAsSingleton = false) : void
     {
-        $binding = new FactoryBinding($factory, $resolveAsSingleton);
+        $binding = new FactoryContainerBinding($factory, $resolveAsSingleton);
 
         foreach ((array)$interfaces as $interface) {
             $this->addBinding($interface, $binding);
@@ -57,7 +57,7 @@ class Container implements IContainer
      */
     public function bindInstance($interfaces, $instance) : void
     {
-        $binding = new InstanceBinding($instance);
+        $binding = new InstanceContainerBinding($instance);
 
         foreach ((array)$interfaces as $interface) {
             $this->addBinding($interface, $binding);
@@ -70,7 +70,7 @@ class Container implements IContainer
     public function bindPrototype($interfaces, string $concreteClass = null, array $primitives = []) : void
     {
         foreach ((array)$interfaces as $interface) {
-            $this->addBinding($interface, new ClassBinding($concreteClass ?? $interface, $primitives, false));
+            $this->addBinding($interface, new ClassContainerBinding($concreteClass ?? $interface, $primitives, false));
         }
     }
 
@@ -80,7 +80,7 @@ class Container implements IContainer
     public function bindSingleton($interfaces, string $concreteClass = null, array $primitives = []) : void
     {
         foreach ((array)$interfaces as $interface) {
-            $this->addBinding($interface, new ClassBinding($concreteClass ?? $interface, $primitives, true));
+            $this->addBinding($interface, new ClassContainerBinding($concreteClass ?? $interface, $primitives, true));
         }
     }
 
@@ -158,18 +158,18 @@ class Container implements IContainer
         }
 
         switch (get_class($binding)) {
-            case InstanceBinding::class:
-                /** @var InstanceBinding $binding */
+            case InstanceContainerBinding::class:
+                /** @var InstanceContainerBinding $binding */
                 return $binding->getInstance();
-            case ClassBinding::class:
-                /** @var ClassBinding $binding */
+            case ClassContainerBinding::class:
+                /** @var ClassContainerBinding $binding */
                 $instance = $this->resolveClass(
                     $binding->getConcreteClass(),
                     $binding->getConstructorPrimitives()
                 );
                 break;
-            case FactoryBinding::class:
-                /** @var FactoryBinding $binding */
+            case FactoryContainerBinding::class:
+                /** @var FactoryContainerBinding $binding */
                 $factory = $binding->getFactory();
                 $instance = $factory();
                 break;
@@ -179,7 +179,7 @@ class Container implements IContainer
 
         if ($binding->resolveAsSingleton()) {
             $this->unbind($interface);
-            $this->addBinding($interface, new InstanceBinding($instance));
+            $this->addBinding($interface, new InstanceContainerBinding($instance));
         }
 
         return $instance;
@@ -213,9 +213,9 @@ class Container implements IContainer
      * Adds a binding to an interface
      *
      * @param string $interface The interface to bind to
-     * @param IBinding $binding The binding to add
+     * @param IContainerBinding $binding The binding to add
      */
-    protected function addBinding(string $interface, IBinding $binding) : void
+    protected function addBinding(string $interface, IContainerBinding $binding) : void
     {
         if (!isset($this->bindings[$this->currentTarget])) {
             $this->bindings[$this->currentTarget] = [];
@@ -228,9 +228,9 @@ class Container implements IContainer
      * Gets a binding for an interface
      *
      * @param string $interface The interface whose binding we want
-     * @return IBinding|null The binding if one exists, otherwise null
+     * @return IContainerBinding|null The binding if one exists, otherwise null
      */
-    protected function getBinding(string $interface) : ?IBinding
+    protected function getBinding(string $interface) : ?IContainerBinding
     {
         // If there's a targeted binding, use it
         if ($this->currentTarget !== self::$emptyTarget && isset($this->bindings[$this->currentTarget][$interface])) {