+++ src/Opulence/Ioc/IContainerBinding.php
@@ -4,13 +4,17 @@
  * Opulence
  *
  * @link      https://www.opulencephp.com
- * @copyright Copyright (C) 2017 David Young
+ * @copyright Copyright (C) 2019 David Young
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
 
 namespace Opulence\Ioc;
 
-interface IBinding
+/**
+ * Defines the interface for container bindings to implement
+ * @internal
+ */
+interface IContainerBinding
 {
     /**
      * Gets whether or not this binding should be resolved as a singleton