+++ src/Opulence/Ioc/InstanceContainerBinding.php
@@ -4,16 +4,17 @@
  * Opulence
  *
  * @link      https://www.opulencephp.com
- * @copyright Copyright (C) 2017 David Young
+ * @copyright Copyright (C) 2019 David Young
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
 
 namespace Opulence\Ioc;
 
 /**
- * Defines an instance binding
+ * Defines an instance container binding
+ * @internal
  */
-class InstanceBinding implements IBinding
+class InstanceContainerBinding implements IContainerBinding
 {
     /** @var object The instance */
     private $instance;